more doxygen changes. Renamed constants.hpp to sensor_constants.hpp
